Add formatting to the invalid package name error messages
This was not straightforward because Click treats error handling from
interactive prompts a bit differently. Instead of using UsageError's
show() method to display the error, it echos the UsageError's message
attribute. Thus, any styles for the error message should be added to
the message string itself, since the show() method will not be invoked.