Move package name validation to a callback function (#49)
* Validate package name via option callback

This ensures that people get an error immediately when using an invalid
package name instead of telling them after they've filled out all the other
prompts.

* Add formatting to the invalid package name error messages

This was not straightforward because Click treats error handling from
interactive prompts a bit differently. Instead of using UsageError's
show() method to display the error, it echos the UsageError's message
attribute. Thus, any styles for the error message should be added to
the message string itself, since the show() method will not be invoked.

* Update Changelog for new package name callback validation