Add nox to the project (#55)
* wip

* Update nox commands to use uv and use pyproject.toml to get python versions

Also upgrades nox and removes the "max" Python constraint
from pyproject.toml.

* Fix mypy errors

These errors emerged as part of the new nox stuff

* Add a nox session to test install via pyproject.toml (i.e., not lockfile)

* Add coverage reports to nox

* uv sync...shhhh!

* Integrate nox test sessions into ci GitHub workflow

* Run nox via uvx instead of pip installing it

* De-couple coverage and test sessions

The original coverage session depended on a test run using the most
recent python version. However, that optimization makes the coverage
session weirder to work with in CI and it's not really worth it.

* Lower test coverage threshold to 90

It was set to 100 to test the failure mode

* Fix GITHUB_OUTPUT

* Fix job name

* Add matrix for testing docs build

* Use trove classifiers when nox calculates supported python versions

* Update CONTRIBUTING.md to reflect the new nox setup

* Add session name for docs build CI job