Specify uv version in pyproject.toml
To strike a balance between pinning an explicit uv version in GitHub
workflows and having to update a bunch of files when there's a new version,
try specifying a uv version in pyproject.toml.

Because the pyproject.toml version is also checked when running uv commands
locally, be a bit flexible rather than require an exact version.